chore: install ImageIO headers
CGImageSource and CGImageDestination are not actually CoreGraphics APIs
on the reference platform; rather, they are part of the ImageIO
framework and should be installed on the system as such.

However, simply transferring these two APIs from OpalGraphics into a new
"OpalImage" subproject is not so easy, since they have developed in many
ways a symbiotic relationship inside the library.

As a workaround, a GNUmakefile.postamble file has been added to
OpalGraphics which takes care of installing ImageIO headers properly.